requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.4.0
alembic==1.18.4
annotated-types==0.7.0
anyio==4.12.1
arq==0.27.0
asyncpg==0.31.0
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
Deprecated==1.3.1
ecdsa==0.19.1
fastapi==0.115.14
frozenlist==1.8.0
greenlet==3.3.2
h11==0.16.0
hiredis==3.3.0
httptools==0.7.1
idna==3.11
limits==5.8.0
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.1
packaging==26.0
propcache==0.4.1
pyasn1==0.6.2
pycparser==3.0
pydantic==2.10.6
pydantic-settings==2.13.1
pydantic_core==2.27.2
PyJWT==2.11.0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
redis==5.0.8
rsa==4.9.1
sentry-sdk==2.54.0
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.48
starlette==0.46.2
structlog==25.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.34.3
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.1
yarl==1.23.0
